feat: logic rewrite to make calcs with bigger nums
This code allows the user to make calculations with bigger numbers by
taking the initial value of the strings from the keypad labeled as
numbers, and concatenating them together within a "valueCurrent" array.
As the operation progress, the logic separates the first number into
its own "valueCurrent" variable, and the second number into a
"valueCurrent1" variable during the process in which a user does the
following:

    1. The user inputs the first variable
    2. The user inputs the arthimetic operation of choice (+, -, etc.)
    3. The user inputs a second variable

The calculator currently auto-resolves the equation this way, and places
the resulting value in the display as well as a "valueCurrent2" variable
while wiping away the previous `valueCurrent*` variables in an attempt
to recreate the cycle.

This solves a lot of problems with the previous implementation, but
doesn't resolve the following:

    1. The first two variables should only auto-resolve if another operator
	is entered OR an equals operator is clicked by the user
    2. The variables currently do not continuously "cycle"/calculate.
    3. Currently, string numbers inputted by the user are turned into a
       radix of ten, but it is ambigious to me right now as to whether
       decimals properly round (they probably don't).
    4. CLEAR has no current functionality
    5. It is ambigious whether attempting to press an extra operator
       between user steps 2 & 3 will result in a crash. This should be
       ensured to be resolved.